+++ inst/app.R
@@ -1,9 +1,0 @@
-library(ambiorix)
-
-app <- Ambiorix$new(log = TRUE)
-
-app$get("/", function(req, res){
-  res$send("Hello {ambiorix}!")
-})
-
-app$start()